--- a/web/src/lib/app.tsx
+++ b/web/src/lib/app.tsx
@@ -121,32 +121,27 @@ export default class App extends React.Component<{}, State> {
   }
 
   render() {
-    const loaded = this.state.loaded;
-    return (
-      <div>
-        {loaded ? (
-          <Router>
-            <Pages
-              user={this.user}
-              api={this.api}
-              match={null}
-              location={null}
-              history={null}
-            />
-          </Router>
-        ) : (
-          <div id="spinner">
-            <span
-              ref={el => {
-                if (this.progressMeter) {
-                  return;
-                }
+    return this.state.loaded ? (
+      <Router>
+        <Pages
+          user={this.user}
+          api={this.api}
+          match={null}
+          location={null}
+          history={null}
+        />
+      </Router>
+    ) : (
+      <div id="spinner">
+        <span
+          ref={el => {
+            if (this.progressMeter) {
+              return;
+            }
 
-                this.progressMeter = el as HTMLSpanElement;
-              }}
-            />
-          </div>
-        )}
+            this.progressMeter = el as HTMLSpanElement;
+          }}
+        />
       </div>
     );
   }