--- a/web/src/lib/components/index.css
+++ b/web/src/lib/components/index.css
@@ -371,11 +371,8 @@ body:not(.ios) .recording #background-container {
 }
 
 #content > div {
-  display: none;
-  opacity: 0;
   box-sizing: border-box;
   transition: opacity 0.4s ease-in;
-  z-index: var(--bottom-z-index);
   max-width: var(--desktop-width);
   margin: 0 auto;
 }
@@ -386,12 +383,6 @@ body:not(.ios) .recording #background-container {
     min-height: 100%;
     padding: 0 var(--page-margin);
   }
-}
-
-#content > div.active {
-  display: block;
-  opacity: 1;
-  z-index: var(--middle-z-index);
 }
 
 .main-logo {
@@ -631,12 +622,6 @@ body:not(.mobile-safari) .robot .bubble:hover {
   background-color: var(--base-background-color);
   max-width: var(--desktop-width);
   padding: var(--page-margin);
-  transform: scale(0);
-  transition: transform 0.2s ease-out;
-}
-
-.active .privacy-content {
-  transform: scale(1);
 }
 
 @media (min-width: 600px) {