--- a/web/src/lib/components/pages/home/home.css
+++ b/web/src/lib/components/pages/home/home.css
@@ -1,4 +1,4 @@
-.home .bubble {
+.home .robot .bubble {
     display: flex;
 }