--- a/web/src/lib/components/pages/listen.tsx
+++ b/web/src/lib/components/pages/listen.tsx
@@ -17,7 +17,6 @@ export default class Listen extends React.Component<ListenPageProps, {}> {
 
   onVote() {
     this.props.user.tallyVerification();
-    // this.props.navigate('/listen'); // force page render
   }
 
   render() {