--- a/web/src/lib/components/pages/not-found.tsx
+++ b/web/src/lib/components/pages/not-found.tsx
@@ -1,10 +1,10 @@
-import { h, Component } from 'preact';
+import * as React from 'react';
 
 interface Props {
   active: string;
 }
 
-export default class NotFound extends Component<Props, void> {
+export default class NotFound extends React.Component<Props, {}> {
   render() {
     return (
       <div id="not-found-container" className={this.props.active}>