… CONTRIBUTING.md
… package.json
… tsconfig.json
… web/index.html
… web/src/lib/components/pages/privacy.tsx
… web/src/lib/components/pages/profile.tsx
… web/src/lib/components/pages/record/profile-actions.tsx
… web/src/lib/components/pages/record/record.tsx
… web/src/lib/components/pages/record/review.tsx
… web/src/lib/components/pages/terms.tsx
… web/src/lib/components/privacy-content.tsx
… web/src/lib/components/profile-form/profile-form.tsx
… web/src/lib/components/progress-button.tsx
… web/src/lib/components/robot-talker.tsx
… web/src/lib/components/robot.tsx
… web/src/lib/components/terms-content.tsx
… web/src/lib/components/validator.tsx
… web/src/main.ts
… web/webpack.prod.config.js
… yarn.lock