## Upsides

- No more navigational logic in our codebase, yay! (but one more dependency 🙈 )
- React is strict about where in the lifecycle setState is called. We had a couple of places that violated that, imo the code got more clear by fixing those. One thing I found that is a React anti-pattern is setting state that can be derived from props.
- `app.tsx` is finally a component :)

## Bundle Size

| | Before | After |
| ------------- | ------------- | ------------- |
| **Uncompressed** | 100kb  | 252kb (**+40%**)   |
| **Compressed** | 27kb | 71kb (**+38%**) |

![screenshot-2017-11-8 webpack bundle analyzer](https://user-images.githubusercontent.com/4051932/32547079-531b4cac-c481-11e7-9120-427ec5181f2d.png)

## Performance

@mikehenrty was also curious about performance, especially how conditional rendering compares to switching visibility via CSS. Now I just took a run at using Chrome's "reload and profile" to see where the first meaningful render occurs. The numbers are averaged over 6 runs each from my T450s with the CPU throttled x6 in Chrome (all local requests). The commit I used for the Preact version is 3ae023f as it was the one right before we started switching to conditional renders in some places.

| | Before | After |
| - | - | - |
| **Initial Render** | 3000ms | 2500ms (**-17%**) |
| **While Recording** | without jank | no jank here either |

Now I'd like to do a more detailed comparison between the two, but unfortunately I did not find a good methodology on how to record the exact same sequence of events. In this case, that would be going through recording sentences and then submitting them. As a result I'd want to see the avg FPS, or more importantly if there was jank.

While we don't have that, we'll have to make it more of a gut decision. Though what I can say with confidence is that I've not yet found a case where it wasn't possible to speed React up and I don't think the speed benefits Preact might have in certain cases should have a large impact on our app. But again, I'd like to back that up with more data, so if anyone has a good methodology to offer, feel free to flame me here or anywhere else 🙂 

## Related issues
- Fixes #591
- Closes #454