@Gregoor 2017-11-07
@kenrick95 here it is. Still in a broken state. I started the migration to `react-router` top-down, meaning I've just looked into `app.tsx` for now (though didn't actually import anything from `react-router` yet) My thinking was we'd put all the routing logic into `app.tsx` and keep `pages.tsx` as a nav & layout file (maybe we'd also rename it to `layout.tsx`). LMKWYT!
I gave you access to push to this branch of my fork, so we have an easier time working on this.