@kenrick95 2017-11-07
Actually, routing at react router v4 is very simple and routing could be implemented at component level (hence it could be kept at `pages.tsx` though it could make the page quite long). I've pushed a commit that more or less similar to previously implemented at #472 . Let me know what you think about the style.

Note: it doesn't compile yet, as there are too many errors when compiling, I'm not sure yet how to fix it.

Thanks