@Gregoor 2017-11-07
Thanks lots, that was quick! I think I'll still move the routing to `app.tsx` because while it is a larger change that way, it makes `pages.tsx` tigther and only layout related. Will ping you when I have it done, if you want to review it :)