@Gregoor 2017-11-07
Okay, I did not move it yet @kenrick95. Think I'll want to wait until we move some code out of `pages.tsx` into their own actions (for example powered by Redux).
___
@mikehenrty This should be on par with the preact version now, feature wise. Imo code quality also improved:
- No more navigational logic in our codebase, yay! (but one more dependency 🙈 )
- React is strict about where in the lifecycle `setState` is called. We had a couple of places that violated that, imo the code got more clear by fixing those. One thing I found that is a React anti-pattern is setting state that can be derived from props.
- `app.tsx` is finally a component :)