@Gregoor 2017-11-08
I just noticed that the bundle size increased by a factor of 4.6 (before 100kb, after 461kb). Investigating that now and will put results here.