@Gregoor 2017-11-08
> When I build locally, I see the bundle size go up to be 3mb uncompressed. Is anyone else seeing this? If so, my opinion is that this is a dealbreaker in terms of size. Maybe there is something I am doing wrong?

You're doing nothing wrong, but this time the dev build is not minified and some libraries include different things depending on environment, in the case of React that's for example helpful warning when breaking implicit contracts (like we did in a couple of places). These messages re not visible (or shortened to an error code) in production. So the production build is actually 250kb and 75kb gzipped (numbers are also in the first post). So it depends on whether you run `start` or `start:prod`.


> But doing so also has noticeable performance penalties due to things like DOM reflow. For instance, reflow on the recording screen has become noticeble on the confirm submission form. That was done in a different commit and so we won't block here, but I'd like us to start taking a more User-centered approach rather than developer centered-approach when developing this codebase.

Thanks, that's a good point, that wasn't on my radar enough. My intuition would say that we paid for the "extra-speed" with longer initial renders, but maybe that's indeed a price worth paying when things run smoother once loaded. I'll start measuring things and see if I can update my intuition with actual numbers.